[PAN-1085] Undoing some changes for simplicity
