+++ aurora/amun/client/utils.py
@@ -101,24 +101,3 @@ def get_scenario_by_name(scenarios, scenario_name: str):
 
 def get_v2_url(url):
     return url.replace("/v1/", "/v2/")
-
-# Sets Up root loging console (INFO) and file (DEBUG) handlers
-def setup_file_and_console_loggers(fileName):
-    os.makedirs("logs", exist_ok=True)
-    rotFileHandler = logging.handlers.RotatingFileHandler(
-        f"logs/{fileName}", "a", 30 * 1024 * 1024, 10
-    )
-    f = logging.Formatter("%(asctime)s %(name)s %(levelname)-8s %(message)s")
-    rotFileHandler.setFormatter(f)
-    rotFileHandler.setLevel(logging.DEBUG)
-    consoleHandler = logging.StreamHandler()
-    consoleHandler.setLevel(logging.INFO)
-    consoleHandler.setFormatter(
-        logging.Formatter("%(asctime)s %(levelname)-8s %(message)s")
-    )
-    logger = logging.getLogger()
-    logger.addHandler(rotFileHandler)
-    logger.addHandler(consoleHandler)
-    log.setLevel(logging.DEBUG)
-    # Set Level for Amun SDK
-    logging.getLogger("aurora.amun").setLevel(logging.DEBUG)
